add textures to format and fix parsing error

+++ BlockbenchToPmodel/Blockbench/Rotation.cs
@@ -4,7 +4,7 @@ namespace BlockbenchToPmodel.Blockbench
 {
     internal class Rotation
     {
-        [JsonProperty("angle")] public int Angle { get; set; }
+        [JsonProperty("angle")] public float Angle { get; set; }
         [JsonProperty("axis")] public string Axis { get; set; }
         [JsonProperty("origin")] public float[] Origin { get; set; }
     }
